Move Spinning component to common folder
